Llamaindex and RAG/requirements.txt
llama_index
llama-index-readers-web
python-dotenv
html2text